+++ src/bika/lims/content/analysisrequest.py
@@ -21,6 +21,7 @@
 import base64
 import functools
 import re
+from datetime import datetime
 from decimal import Decimal
 
 from AccessControl import ClassSecurityInfo
@@ -429,7 +430,7 @@ schema = BikaSchema.copy() + Schema((
     DateTimeField(
         'DateSampled',
         mode="rw",
-        max="created",
+        max="getMaxDateSampled",
         read_permission=View,
         write_permission=FieldEditDateSampled,
         widget=DateTimeWidget(
@@ -2593,6 +2594,14 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
         if num_steps > max_num_steps:
             return 100
         return (num_steps * 100) / max_num_steps
+
+    def getMaxDateSampled(self):
+        """Returns the maximum date for sample collection
+        """
+        if not self.getSamplingWorkflowEnabled():
+            # no future, has to be collected before registration
+            return api.get_creation_date(self)
+        return datetime.max
 
 
 registerType(AnalysisRequest, PROJECTNAME)