+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.5.0 (unreleased)
 ------------------
 
+- #2402 Fix user cannot enter future date for DateSampled when sampling enabled
 - #2401 Fix OverflowError when calculating datetime.min date for left-hand TZs
 - #2399 Support for min/max in DateTimeWidget, and drop _nopast and _nofuture
 - #2397 Fix district is not displayed in old address widget